fix(routes): resolve 404/403/500 errors on analytics, insights, audit, permissions
Routes & Templates:
- Change /finance/insights to /insights route
- Change /finance/audit/timeline to /audit route
- Create permissions/list.html template
- Update form actions and export links in templates

Analytics/Insights SQL:
- Add explicit ::bigint casts to nullable branch_id parameters
- Fix 'could not determine data type of parameter' PostgreSQL errors
- Update both source SQL and generated Go files

Permissions:
- Add migration 000015 for finance analytics permissions
- Add finance.view_insights, finance.view_audit, permissions.view to seed
- Assign new permissions to admin role

Sidebar:
- Fix double-active state on /analytics/kpi
- Use exact match only for nav-item active class

Workflow:
- Add fix-backend.md workflow for backend debugging

+++ internal/analytics/db/analytics.sql.go
@@ -14,7 +14,7 @@ SELECT a.bucket, a.amount::double precision AS amount, a.as_of
 FROM mv_ap_aging a
 WHERE a.as_of = $1
   AND a.company_id = $2
-  AND ($3 IS NULL OR a.branch_id = $3)
+  AND ($3::bigint IS NULL OR a.branch_id = $3::bigint)
 ORDER BY a.bucket
 `
 
@@ -55,7 +55,7 @@ SELECT a.bucket, a.amount::double precision AS amount, a.as_of
 FROM mv_ar_aging a
 WHERE a.as_of = $1
   AND a.company_id = $2
-  AND ($3 IS NULL OR a.branch_id = $3)
+  AND ($3::bigint IS NULL OR a.branch_id = $3::bigint)
 ORDER BY a.bucket
 `
 
@@ -101,7 +101,7 @@ WITH pl AS (
   FROM mv_pl_monthly m
   WHERE m.period = $1
     AND m.company_id = $2
-    AND ($3 IS NULL OR m.branch_id = $3)
+    AND ($3::bigint IS NULL OR m.branch_id = $3::bigint)
 ), cash AS (
   SELECT
     COALESCE(SUM(m.cash_in)::double precision, 0) AS cash_in,
@@ -109,19 +109,19 @@ WITH pl AS (
   FROM mv_cashflow_monthly m
   WHERE m.period = $1
     AND m.company_id = $2
-    AND ($3 IS NULL OR m.branch_id = $3)
+    AND ($3::bigint IS NULL OR m.branch_id = $3::bigint)
 ), ar AS (
   SELECT COALESCE(SUM(a.amount)::double precision, 0) AS outstanding
   FROM mv_ar_aging a
   WHERE a.as_of = $4
     AND a.company_id = $2
-    AND ($3 IS NULL OR a.branch_id = $3)
+    AND ($3::bigint IS NULL OR a.branch_id = $3::bigint)
 ), ap AS (
   SELECT COALESCE(SUM(a.amount)::double precision, 0) AS outstanding
   FROM mv_ap_aging a
   WHERE a.as_of = $4
     AND a.company_id = $2
-    AND ($3 IS NULL OR a.branch_id = $3)
+    AND ($3::bigint IS NULL OR a.branch_id = $3::bigint)
 )
 SELECT
   pl.net_profit,
@@ -186,7 +186,7 @@ SELECT m.period,
 FROM mv_cashflow_monthly m
 WHERE m.period BETWEEN $1 AND $2
   AND m.company_id = $3
-  AND ($4 IS NULL OR m.branch_id = $4)
+  AND ($4::bigint IS NULL OR m.branch_id = $4::bigint)
 ORDER BY m.period
 `
 
@@ -247,7 +247,7 @@ SELECT m.period,
 FROM mv_pl_monthly m
 WHERE m.period BETWEEN $1 AND $2
   AND m.company_id = $3
-  AND ($4 IS NULL OR m.branch_id = $4)
+  AND ($4::bigint IS NULL OR m.branch_id = $4::bigint)
 ORDER BY m.period
 `
